requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
elevenlabs>=0.3.0
python-dotenv>=1.0.0
# sounddevice>=0.4.6
# scipy>=1.11.0
# numpy>=1.24.0
openai-agents>=0.0.13
openai>=1.76.0
requests
Unidecode
# Used for audio transcription
pydub==0.25.1

# Document processing
PyPDF2==3.0.1
python-docx==1.0.1
langchain==0.1.9
langchain-community==0.0.29
faiss-cpu==1.7.4
sentence-transformers==2.5.1
numpy==1.26.4